requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
regex
google-generativeai
